app/requirements.txt
boto3==1.34.0
botocore==1.34.0
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
Pillow==10.1.0
pandas==2.1.4
numpy==1.25.2
werkzeug==3.0.1
gunicorn==21.2.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-flask==0.42b0
opentelemetry-instrumentation-requests==0.42b0
opentelemetry-instrumentation-boto3sqs==0.42b0
opentelemetry-instrumentation-logging==0.42b0
opentelemetry-exporter-prometheus==1.12.0rc1
opentelemetry-exporter-jaeger==1.17.0
opentelemetry-exporter-otlp==1.21.0
prometheus-client==0.19.0
